Revising logic for formatting estimates and adding getProviders() in estimateContainer's mount function
