+++ eventkit_cloud/ui/static/ui/app/components/CreateDataPack/DataProvider.tsx
@@ -235,14 +235,6 @@ export class DataProvider extends React.Component<Props, State> {
             const estimateInSeconds = providerEstimates.time.value;
             durationEstimate = getDuration(estimateInSeconds);
         }
-
-        // if (providerEstimates.estimates.size) {
-        //     sizeEstimate = formatMegaBytes(providerEstimates.size.value);
-        // }
-        // if (providerEstimates.estimates.time) {
-        //     const estimateInSeconds = providerEstimates.time.value;
-        //     durationEstimate = getDuration(estimateInSeconds);
-        // }
         return `${get(sizeEstimate)} / ${get(durationEstimate)}`;
     }
