+++ eventkit_cloud/ui/static/ui/app/styles/ol3map.css
@@ -3,7 +3,6 @@ div.map {
     margin-top: 0px;
     margin-left: 0px;
     bottom: 0px;
-    /*height: calc(100% - 180px);*/
     height: calc(100% - 210px);
 }
